+++ modules/Attendance/TeacherCompletion.php
@@ -95,7 +95,8 @@ if ( SchoolInfo( 'NUMBER_DAYS_ROTATION' ) !== null )
 				AND SCHOOL_ID=acc.SCHOOL_ID
 				AND SYEAR=acc.SYEAR)
 			AND CALENDAR_ID=cp.CALENDAR_ID)
-		AS INT) FOR 1) IN cpsp.DAYS)>0 OR (sp.BLOCK IS NOT NULL AND sp.BLOCK=acc.BLOCK))
+		" . ( $DatabaseType === 'mysql' ? "AS UNSIGNED)" : "AS INT)" ) .
+		" FOR 1) IN cpsp.DAYS)>0 OR (sp.BLOCK IS NOT NULL AND sp.BLOCK=acc.BLOCK))
 	ORDER BY FULL_NAME";
 }
 else